LayoutTests/imported/mozilla/css-transitions/test_effect-target-expected.txt


PASS Returned CSS transitions have the correct Animation.target 
PASS effect.target should return the same CSSPseudoElement object each time 
FAIL effect.target from the script-generated animation should return the same CSSPseudoElement object as that from the CSS generated transition assert_equals: Got animations running on ::after pseudo element expected 2 but got 4
